+++ src/main/java/com/minhcraft/createlesstrains/CreatelessTrains.java
@@ -1,9 +1,7 @@
 package com.minhcraft.createlesstrains;
 
+import com.minhcraft.createlesstrains.feature.automatic_door.VanillaDoorMovementBehaviorRegistry;
 import com.simibubi.create.CreateBuildInfo;
-
-import com.simibubi.create.api.behaviour.movement.MovementBehaviour;
-import com.simibubi.create.content.decoration.slidingDoor.SlidingDoorMovementBehaviour;
 
 import io.github.fabricators_of_create.porting_lib.util.EnvExecutor;
 import net.fabricmc.api.ModInitializer;
@@ -12,8 +10,6 @@ import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.resources.ResourceLocation;
-
-import net.minecraft.world.level.block.Blocks;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,32 +34,7 @@ public class CreatelessTrains implements ModInitializer {
 						modContainer ->
 								ResourceManagerHelper.registerBuiltinResourcePack(CreatelessTrains.id(MOD_ID), modContainer, ResourcePackActivationType.ALWAYS_ENABLED));
 
-
-		// Make it so vanilla doors open and close automatically on trains, just like create's train doors and framed glass doors
-		MovementBehaviour.REGISTRY.register(
-				Blocks.OAK_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.IRON_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.SPRUCE_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.BIRCH_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.JUNGLE_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.DARK_OAK_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.CRIMSON_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.ACACIA_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.BAMBOO_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.CHERRY_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.MANGROVE_DOOR, new SlidingDoorMovementBehaviour());
-		MovementBehaviour.REGISTRY.register(
-				Blocks.WARPED_DOOR, new SlidingDoorMovementBehaviour());
+		VanillaDoorMovementBehaviorRegistry.init();
 	}
 
 	public static ResourceLocation id(String path) {